Narrow PR scope to RoboCup team communication UDP side channel only
- Revert the Overview module changes (module/output/Overview) back to
  main, dropping the Overview UDP side channel feature entirely.
- Remove OverviewUDPClient and the generic UDPSideChannelClient base
  class; RoboCupUDPClient is now a single self-contained client that
  delivers raw message.input.RoboCup packets directly (rather than
  translating them into an Overview packet).
- Add a RoboCup network handler in the Localisation view that maps
  the fields RoboCup shares with Overview (pose, ball, kick target,
  walk command) onto the same DashboardRobotModel used for Overview,
  so team communication packets are visualised without depending on
  the Overview message format.
- Drop the overview-udp-peer peer type and --overview-udp-* CLI args;
  only --robocup-udp-port/--robocup-udp-allowed-addresses remain.
