+++ nusight2/src/server/nusight_server.ts
@@ -6,7 +6,6 @@ import { NUClearNetClient } from "../shared/nuclearnet/nuclearnet_client";
 import { CompositeNUClearNetClient } from "./nuclearnet/composite_nuclearnet_client";
 import { DirectNUClearNetClient } from "./nuclearnet/direct_nuclearnet_client";
 import { FakeNUClearNetClient } from "./nuclearnet/fake_nuclearnet_client";
-import { OverviewUDPClient, OverviewUDPClientOptions } from "./nuclearnet/overview_udp_client";
 import { RoboCupUDPClient, RoboCupUDPClientOptions } from "./nuclearnet/robocup_udp_client";
 import { NUsightSession } from "./session/session";
 import { ClientConnection } from "./web_socket/client_connection";
@@ -17,14 +16,8 @@ interface NUsightServerOpts {
   connectionOpts: NUClearNetOptions;
   /**
    * When provided, an additional receive-only UDP side channel is opened that presents robots
-   * sending serialised Overview packets as their own peers, separate from the main NUClearNet
-   * connection.
-   */
-  overviewUDP?: OverviewUDPClientOptions;
-  /**
-   * When provided, an additional receive-only UDP side channel is opened that presents robots
    * (or other teams) sending serialised RoboCup team communication packets as their own peers,
-   * translating the fields they share with Overview so they can be displayed the same way.
+   * separate from the main NUClearNet connection.
    */
   robocupUDP?: RoboCupUDPClientOptions;
 }
@@ -69,25 +62,15 @@ export class NUsightServer {
     ]);
   }
 
-  static of(
-    server: WebSocketServer,
-    { fakeNetworking, connectionOpts, overviewUDP, robocupUDP }: NUsightServerOpts,
-  ): NUsightServer {
+  static of(server: WebSocketServer, { fakeNetworking, connectionOpts, robocupUDP }: NUsightServerOpts): NUsightServer {
     const primaryClient: NUClearNetClient = fakeNetworking ? FakeNUClearNetClient.of() : DirectNUClearNetClient.of();
 
-    // Optionally add the UDP side channels as additional receive-only sources. Each is presented
-    // to clients as its own peer (see overview_udp_client.ts and robocup_udp_client.ts), separate
-    // from the main NUClearNet peers.
-    const secondaryClients: NUClearNetClient[] = [];
-    if (overviewUDP) {
-      secondaryClients.push(OverviewUDPClient.of(overviewUDP));
-    }
-    if (robocupUDP) {
-      secondaryClients.push(RoboCupUDPClient.of(robocupUDP));
-    }
-
-    const nuclearnetClient: NUClearNetClient =
-      secondaryClients.length > 0 ? CompositeNUClearNetClient.of(primaryClient, secondaryClients) : primaryClient;
+    // Optionally add the RoboCup UDP side channel as an additional receive-only source. It is
+    // presented to clients as its own peer (see robocup_udp_client.ts), separate from the main
+    // NUClearNet peers.
+    const nuclearnetClient: NUClearNetClient = robocupUDP
+      ? CompositeNUClearNetClient.of(primaryClient, [RoboCupUDPClient.of(robocupUDP)])
+      : primaryClient;
 
     return new NUsightServer(server, nuclearnetClient, connectionOpts);
   }
